drm-howto: add modesetting example and howto
The modeset.c example is a short tutorial how to use the DRM mode-setting
API. It describes the concept of connectors, CRTCs, encoders and
framebuffers and shows how to make use of them in simple
graphics-applications.

It is targeted at users with no prior understandings of the DRM system and
is a requirement for understanding any more complex DRM tutorials.

Signed-off-by: David Herrmann <dh.herrmann@googlemail.com>