allow specifying the mode in the command line
With this change, the output from "cvt x y clock" can be used as parameters for
modeset.c. This allows for modes that are not among the modes of a connector.
For example, 800x400 can be be used even if it is not one of the modes of the
active connectors:

modeset $(cvt 800 400 75 | grep -v '^#')

Of course, this may fail if not supported by the hardware.

Two other changes are made necessary by this:

1. instead of 'modeset /dev/dri/card0', this parameter now requires -d, as
   otherwise the program would not be able to tell this from the name of the
   mode

2. filling the screen with the same color hides the difference in resolution;
   this is now enphasized by a small black 10x10 square at position 100,100