drm-howto: add atomic API example
Most modern drivers make use of the KMS atomic API, since it fix
problems related to synchronisation. So it's important to have
an example that uses the atomic API.

@ezequielgarcia wrote the code to extend the double-buffered
vsync'ed example to use the atomic API. I've added comments to
it and with the help of @pq added some changes to the code.

Signed-off-by: Ezequiel Garcia <ezequiel@collabora.com>
Signed-off-by: Leandro Ribeiro <leandro.ribeiro@collabora.com>